Add more notification actions and multiple shield actions
