+++ packages/react-native-device-activity/ios/Shared.swift
@@ -97,6 +97,126 @@ func notifyAppWithName(name: String) {
   CFNotificationCenterPostNotification(notificationCenter, notificationName, nil, nil, false)
 }
 
+func executeGenericAction(action: [String: Any], placeholders: [String: String?], eventKey: String) {
+  let type = action["type"] as? String
+
+  if let sleepBefore = action["sleepBefore"] as? Int {
+    sleep(ms: sleepBefore)
+  }
+
+  if type == "blockSelection" {
+    if let familyActivitySelectionId = action["familyActivitySelectionId"] as? String {
+      if let activitySelection = getFamilyActivitySelectionById(id: familyActivitySelectionId) {
+        updateShield(
+          shieldId: action["shieldId"] as? String,
+          triggeredBy: eventKey,
+          activitySelectionId: familyActivitySelectionId
+        )
+
+        sleep(ms: 50)
+
+        blockSelectedApps(
+          blockSelection: activitySelection,
+          triggeredBy: eventKey
+        )
+      } else {
+        logger.log("No familyActivitySelection found with ID: \(familyActivitySelectionId)")
+      }
+    }
+  } else if type == "unblockSelection" {
+    if let familyActivitySelectionId = action["familyActivitySelectionId"] as? String {
+      if let activitySelection = getFamilyActivitySelectionById(id: familyActivitySelectionId) {
+
+        unblockSelection(
+          removeSelection: activitySelection,
+          triggeredBy: eventKey
+        )
+
+        userDefaults?
+          .removeObject(
+            forKey: SHIELD_CONFIGURATION_FOR_SELECTION_PREFIX + "_" + familyActivitySelectionId)
+      }
+    }
+  } else if type == "addSelectionToWhitelist" {
+    if let familyActivitySelectionInput = action["familyActivitySelection"] as? [String: Any] {
+      let selection = parseActivitySelectionInput(input: familyActivitySelectionInput)
+      addSelectionToWhitelistAndUpdateBlock(
+        whitelistSelection: selection,
+        triggeredBy: eventKey
+      )
+    }
+  } else if type == "removeSelectionFromWhitelist" {
+    if let familyActivitySelectionInput = action["familyActivitySelection"] as? [String: Any] {
+      let selection = parseActivitySelectionInput(input: familyActivitySelectionInput)
+      removeSelectionFromWhitelistAndUpdateBlock(
+        selection: selection,
+        triggeredBy: eventKey
+      )
+    }
+  } else if type == "clearWhitelistAndUpdateBlock" {
+    logger.info("should clearWhitelistAndUpdateBlock")
+    clearWhitelist()
+    updateBlock(triggeredBy: eventKey)
+    logger.info("done")
+  } else if type == "resetBlocks" {
+    resetBlocks(triggeredBy: eventKey)
+  } else if type == "clearWhitelist" {
+    clearWhitelist()
+  } else if type == "disableBlockAllMode" {
+    disableBlockAllMode(triggeredBy: eventKey)
+  } else if type == "openApp" {
+    // todo: replace with general string
+    openUrl(urlString: "device-activity://")
+
+    sleep(ms: 1000)
+  } else if type == "enableBlockAllMode" {
+    updateShield(
+      shieldId: action["shieldId"] as? String,
+      triggeredBy: eventKey,
+      activitySelectionId: nil
+    )
+
+    // sometimes the shield doesn't pick up the shield config change above, trying a sleep to get around it
+    sleep(ms: 50)
+
+    enableBlockAllMode(triggeredBy: eventKey)
+  } else if type == "removeAllPendingNotificationRequests" {
+    UNUserNotificationCenter.current().removeAllPendingNotificationRequests()
+  } else if type == "removePendingNotificationRequests" {
+    if let identifiers = action["identifiers"] as? [String] {
+      UNUserNotificationCenter
+        .current()
+        .removePendingNotificationRequests(withIdentifiers: identifiers)
+    }
+  } else if type == "setBadgeCount" {
+    let actionWithReplacedPlaceholders = replacePlaceholdersInObject(action, with: placeholders)
+    if let count = actionWithReplacedPlaceholders["count"] as? Int {
+      if #available(iOS 16.0, *) {
+        UNUserNotificationCenter
+          .current()
+          .setBadgeCount(count)
+      }
+    }
+  } else if type == "sendNotification" {
+    if let notification = action["payload"] as? [String: Any] {
+      sendNotification(contents: notification, placeholders: placeholders)
+    }
+  } else if type == "sendHttpRequest" {
+    if let url = action["url"] as? String {
+      let config = action["options"] as? [String: Any] ?? [:]
+
+      task = sendHttpRequest(with: url, config: config, placeholders: placeholders)
+
+      // required for it to have time to trigger before process/callback ends
+      sleep(ms: 1000)
+    }
+  }
+
+  if let sleepAfter = action["sleepAfter"] as? Int {
+    sleep(ms: sleepAfter)
+  }
+}
+
 func sendNotification(contents: [String: Any], placeholders: [String: String?]) {
   let content = UNMutableNotificationContent()
